merge-1003753

See merge request codedriver/codedriver-process!452

+++ src/main/java/codedriver/module/process/api/commenttemplate/ProcessCommentTemplateDeleteApi.java
@@ -1,8 +1,10 @@
 package codedriver.module.process.api.commenttemplate;
 
-import codedriver.framework.auth.core.AuthAction;
+import codedriver.framework.auth.core.AuthActionChecker;
 import codedriver.framework.common.constvalue.ApiParamType;
+import codedriver.framework.exception.type.PermissionDeniedException;
 import codedriver.framework.process.dao.mapper.ProcessCommentTemplateMapper;
+import codedriver.framework.process.dto.ProcessCommentTemplateVo;
 import codedriver.framework.process.exception.commenttemplate.ProcessCommentTemplateNotFoundException;
 import codedriver.framework.reminder.core.OperationTypeEnum;
 import codedriver.framework.restful.annotation.Input;
@@ -10,12 +12,12 @@ import codedriver.framework.restful.annotation.OperationType;
 import codedriver.framework.restful.annotation.Output;
 import codedriver.framework.restful.annotation.Param;
 import codedriver.framework.restful.core.privateapi.PrivateApiComponentBase;
+import codedriver.module.process.auth.label.PROCESS_COMMENT_TEMPLATE_MODIFY;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@AuthAction(name = "PROCESS_COMMENT_TEMPLATE_MODIFY")
 @Service
 @Transactional
 @OperationType(type = OperationTypeEnum.DELETE)
@@ -31,7 +33,7 @@ public class ProcessCommentTemplateDeleteApi extends PrivateApiComponentBase {
 
     @Override
     public String getName() {
-        return "删除系统回复模版";
+        return "删除回复模版";
     }
 
     @Override
@@ -46,6 +48,11 @@ public class ProcessCommentTemplateDeleteApi extends PrivateApiComponentBase {
         Long id = jsonObj.getLong("id");
         if(commentTemplateMapper.checkTemplateExistsById(id) == 0){
             throw new ProcessCommentTemplateNotFoundException(id);
+        }
+        ProcessCommentTemplateVo vo = commentTemplateMapper.getTemplateById(id);
+        /** 没有权限则不允许删除系统模版 */
+        if(ProcessCommentTemplateVo.TempalteType.SYSTEM.getValue().equals(vo.getType()) && !AuthActionChecker.check(PROCESS_COMMENT_TEMPLATE_MODIFY.class.getSimpleName())){
+            throw new PermissionDeniedException();
         }
         commentTemplateMapper.deleteTemplate(id);
         commentTemplateMapper.deleteTemplateAuthority(id);
